One more round of portability fixes for printf'ing size_ts

+++ test/b40c_util.h
@@ -274,7 +274,7 @@ int CompareResults(T* computed, T* reference, size_t len, bool verbose)
 	for (size_t i = 0; i < len; i++) {
 
 		if (computed[i] != reference[i]) {
-			printf("Incorrect: [%zu]: ", i);
+			printf("Incorrect: [%lu]: ", (unsigned long) i);
 			PrintValue<T>(computed[i]);
 			printf(" != ");
 			PrintValue<T>(reference[i]);
