+++ test/scan/tune_scan.cu
@@ -312,7 +312,7 @@ public:
 	template <typename ScanConfig>
 	void TimedScan()
 	{
-		printf("%zu, ", sizeof(T));
+		printf("%lu, ", (unsigned long) sizeof(T));
 		ScanConfig::Print();
 		fflush(stdout);
 
@@ -570,7 +570,7 @@ int main(int argc, char** argv)
 
 	util::CudaProperties cuda_props;
 
-	printf("Test Scan: %d iterations, %zu elements", g_iterations, num_elements);
+	printf("Test Scan: %d iterations, %lu elements", g_iterations, (unsigned long) num_elements);
 	printf("\nCodeGen: \t[device_sm_version: %d, kernel_ptx_version: %d]\n\n",
 		cuda_props.device_sm_version, cuda_props.kernel_ptx_version);
